fix fixtures user
