+++ src/DataFixtures/CategoryFixtures.php
@@ -23,10 +23,11 @@ class CategoryFixtures extends Fixture
 
     public function load(ObjectManager $manager): void
     {
-        foreach (self::CATEGORIES as $categoryName) {
+        foreach (self::CATEGORIES as $index => $categoryName) {
             $category = new Category();
             $category->setName($categoryName);
             $manager->persist($category);
+            $this->addReference('category_' . $index, $category);
         }
 
         $manager->flush();
